fix: move LeafletToolsTests to correct MCP/Tools location and add LogError assertion
- Moved LeafletToolsTests from Features/Leaflet/MCP/ to MCP/Tools/ to match existing MCP tool test conventions
- Updated namespace from Anela.Heblo.Tests.Features.Leaflet.MCP to Anela.Heblo.Tests.MCP.Tools
- Added LogLevel.Error verification to GenerateLeaflet_wraps_unexpected_exception_with_generic_message test
- Removed trailing comma from GenerateLeafletRequest object initializer in LeafletTools.cs
